+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Layouts
 
+## Main
+
+- Add `ErrorBoundary` component. [#272](https://github.com/mapbox/dr-ui/pull/272)
+
 ## 0.29.2
 
 - Add aria-label to `BackToTopButton`. [#290](https://github.com/mapbox/dr-ui/pull/290)